+++ client/models/global_http_event_collector.go
@@ -14,11 +14,11 @@ type GlobalHECEntry struct {
 type GlobalHttpEventCollectorObject struct {
 	Host                string `json:"host,omitempty" url:"host,omitempty"`
 	Index               string `json:"index,omitempty" url:"index,omitempty"`
-	Disabled            bool   `json:"disabled,string,omitempty" url:"disabled"`
+	Disabled            bool   `json:"disabled,omitempty" url:"disabled"`
 	EnableSSL           bool   `json:"enableSSL,string,omitempty" url:"enableSSL"`
 	Port                int    `json:"port,string,omitempty" url:"port,omitempty"`
 	DedicatedIoThreads  int    `json:"dedicatedIoThreads,string,omitempty" url:"dedicatedIoThreads,omitempty"`
 	MaxSockets          int    `json:"maxSockets,string,omitempty" url:"maxSockets,omitempty"`
 	MaxThreads          int    `json:"maxThreads,string,omitempty" url:"maxThreads,omitempty"`
-	UseDeploymentServer bool   `json:"useDeploymentServer,string,omitempty" url:"useDeploymentServer"`
+	UseDeploymentServer int    `json:"useDeploymentServer,string,omitempty" url:"useDeploymentServer"`
 }