+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0 (September 25, 2020)
+* Change HTTP inputs resource attribute types (`use_ack` and `use_deployment_server`) to `int`
+* Creating examples folder
+* Adding AWS firehose + Splunk integration example under `examples/aws`
+
 ## 1.0.2 (September 21, 2020)
 * Adding Github actions workflow